+++ src/collective/compoundcriterion/tests/adapter.py
@@ -19,3 +19,13 @@ class SampleCompoundCrietrionFilterAdapter(object):
     @property
     def query(self):
         return {'Title': {'query': u'title_with_sample_text'}}
+
+
+class WrongFormatCompoundCrietrionFilterAdapter(object):
+
+    def __init__(self, context):
+        self.context = context
+
+    @property
+    def query(self):
+        return {'portal_type': ['Document', 'Folder']}