+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Raise a KeyError if the format of the query returned by the named adapter
+  is not compliant with what is returned by
+  plone.app.querystring.queryparser.parseFormquery, this way it behaves
+  correctly with collective.eeafaceted.collectionwidget.
+  [gbastien]
 
 
 0.1 (2015-06-02)
@@ -13,4 +17,3 @@ Changelog
 
 - Initial release.
   [IMIO]
-